CA/requirements.txt
bleach==4.1.0
boto3==1.20.24
botocore==1.23.24
build==0.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
cryptography==36.0.1
Django==2.1
django-storages==1.7
docutils==0.18.1
idna==3.3
importlib-metadata==4.9.0
jeepney==0.7.1
jmespath==0.10.0
keyring==23.4.0
packaging==21.3
pep517==0.12.0
Pillow==8.4.0
pkginfo==1.8.2
pycparser==2.21
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
readme-renderer==32.0
rekognition-jsonParser-MatyasGreff==0.0.1
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
s3transfer==0.5.0
SecretStorage==3.3.1
six==1.16.0
tomli==2.0.0
tqdm==4.62.3
twine==3.7.1
typing_extensions==4.0.1
urllib3==1.26.7
webencodings==0.5.1
zipp==3.6.0
